requirements.txt
Django==3.2.18
django-tree==0.4.1
djangorestframework==3.13.1
Pillow==9.4.0
numpy==1.24.2
pandas==1.5.3
psycopg2-binary==2.9.5
